#1173 refactor renderEntry
